requirements.txt
fastapi==0.109.1
google_api_python_client
google_auth_oauthlib
loguru==0.7.2
numpy==1.24.3
pandas==1.5.3
PyMySQL==1.0.2
python_dateutil==2.8.2
pytz==2024.1
requests
todoist_api_python==3.1.0
uvicorn==0.15.0
python-multipart==0.0.9
SpeechRecognition
quarter_lib==0.0.37
SQLAlchemy==2.0.17
python-telegram-bot==21.4
croniter==2.0.7
msal==1.31.0
markdown==3.7
beautifulsoup4==4.12.3
pygithub==1.55
